move mouse handlers to layer
